third_party/blink/web_tests/fast/scrolling/events/scrollend-event-fired-to-document.html

<!DOCTYPE HTML>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/gesture-util.js"></script>
<script src="../../../virtual/percent-based-scrolling/resources/percent-based-util.js"></script>
<style>
#targetDiv {
  width: 200px;
  height: 200px;
  overflow: scroll;
}

#innerDiv {
  width: 400px;
  height: 400px;
}
</style>

<body style="margin:0">
<div id="targetDiv">
  <div id="innerDiv">
  </div>
</div>
</body>

<script>
if (window.internals)
  internals.runtimeFlags.overscrollCustomizationEnabled = true;

var target_div = document.getElementById('targetDiv');
var rect = target_div.getBoundingClientRect();
var x = rect.left + rect.width / 2;
var y = rect.top + rect.height / 2;

const pixelsToScroll = 800;
const { x: pixelsToScrollX, y: pixelsToScrollY } = calculatePixelsToScroll(
  target_div, pixelsToScroll, pixelsToScroll
);

async function test_scrollend_event_fire(t, source_device,
                                         precise_scrolling_delta) {
  await waitForCompositorCommit();

  // Make sure that no scrollend event is sent to target_div.
  const listener = target_div.addEventListener("scrollend",
      t.unreached_func("target_div got unexpected scrollend event."));
  t.add_cleanup(() => target_div.removeEventListener('scrollend', listener));

  // Scroll left on target div and wait for the doc to get scrollend event.
  let scrollendPromise = waitForScrollendEvent(document);
  await smoothScroll(
    pixelsToScrollX, x, y, source_device, "left", SPEED_INSTANT,
    precise_scrolling_delta
  );
  await scrollendPromise.then((event) => {
    assert_false(event.cancelable);
    // scrollend events are bubbled when the target node is document.
    assert_true(event.bubbles);
  });
  assert_equals(target_div.scrollLeft, 0);

  // Scroll up on target div and wait for the doc to get scrollend event.
  scrollendPromise = waitForScrollendEvent(document);
  await smoothScroll(
    pixelsToScrollY, x, y, source_device, "up", SPEED_INSTANT,
    precise_scrolling_delta
  );
  await scrollendPromise.then((event) => {
    assert_false(event.cancelable);
    // scrollend events are bubbled when the target node is document.
    assert_true(event.bubbles);
  });
  assert_equals(target_div.scrollTop, 0);

  // Wait two additional frames to ensure that unexpected events are not
  // received.
  await raf();
  await raf();
}

promise_test (async (t) => {
  return test_scrollend_event_fire(t, GestureSourceType.TOUCH_INPUT);
}, 'Tests that the document gets scrollend event when no element scrolls by ' +
   'touch.');

promise_test (async (t) => {
  return test_scrollend_event_fire(t, GestureSourceType.MOUSE_INPUT,
      true /* precise_scrolling_delta */);
}, 'Tests that the document gets scrollend event when no element scrolls by '+
   ' wheel in non-animated scrolling path.');

promise_test (async (t) => {
  return test_scrollend_event_fire(t, GestureSourceType.MOUSE_INPUT,
      false /* precise_scrolling_delta */);
}, 'Tests that the document gets scrollend event when no element scrolls by '+
   ' wheel in smooth scrolling path.');
</script>
